\name{MS_STX}
\alias{MS_STX}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Description of floristic ecology
}
\description{
Description of floristic ecology,statistical analysis
}
\usage{
MS_STX(sp)
}

\arguments{
  \item{sp}{
sp information data.fram
}
}

\examples{
data(plantcom)
sp<-plantcom$sp
MS_STX(sp)
  }
