\name{TABLE_ZH}
\alias{TABLE_ZH}

\title{
TABLE of comprehensive flora Statistics
}
\description{TABLE divide by life-style lists(CP),family,life form(water)
}
\usage{
TABLE_ZH(cp)
}

\arguments{
  \item{cp}{
include cp divide information data.fram
}
}

\examples{
data(plantcom)
cp<-CPHF(plantcom$sp)
TABLE_ZH(cp)
}

