\name{TH}
\alias{TH}

\title{
TH
}
\description{

Replace the first occurrence string with SP $TAXA_NAME (ld) according to the SP information table

}
\usage{
TH(sp,text)
}

\arguments{
  \item{sp}{
  sp data.fram(species information)
}
  \item{text}{
  text ,character,you wangto replace to ld)
}

}

\examples{

}
