\name{leplant-package}
\alias{leplant}

\title{
leplant
}
\description{

Description of community characteristics, calculation of diversity and division of groups
}

\details{

\Version: 0.1.0
\Author: le Zhu
\Maintainer: le Zhu <zhulezl@sina.com>

\License: GPL-3
\Imports: stringr, ggplot2,reshape2,tidyverse,plyr
\Encoding: UTF-8

\Built: R 3.5.1; ; 2019-01-30 08:34:58 UTC; windows
}
